+++ desktop/cmp/dash/canvas/DashCanvasViewModel.js
@@ -4,7 +4,9 @@
  *
  * Copyright © 2021 Extremely Heavy Industries Inc.
  */
+import {managed} from '@xh/hoist/core';
 import {DashViewModel} from '@xh/hoist/desktop/cmp/dash/DashViewModel';
+import {FullScreenSupportModel} from '@xh/hoist/desktop/cmp/fullscreenhandler/FullScreenSupportModel';
 import {createObservableRef} from '@xh/hoist/utils/react';
 import {makeObservable, observable} from 'mobx';
 
@@ -25,6 +27,8 @@ export class DashCanvasViewModel extends DashViewModel {
     @observable hideMenuButton;
     /** @member {boolean} */
     @observable hideFullScreenButton;
+    /** @member {FullScreenSupportModel} */
+    @managed fullScreenSupportModel = new FullScreenSupportModel();
 
     constructor(cfg) {
         super(cfg);