+++ desktop/cmp/dock/impl/DockView.js
@@ -8,8 +8,8 @@ import {DockViewModel} from '@xh/hoist/cmp/dock';
 import {div, filler, hbox, span, vbox} from '@xh/hoist/cmp/layout';
 import {hoistCmp, refreshContextView, RenderMode, uses} from '@xh/hoist/core';
 import {button} from '@xh/hoist/desktop/cmp/button';
+import {fullScreenSupport} from '@xh/hoist/desktop/cmp/fullscreenhandler/FullScreenSupport';
 import {Icon} from '@xh/hoist/icon';
-import {dialog} from '@xh/hoist/kit/blueprint';
 import {elementFromContent} from '@xh/hoist/utils/react';
 import classNames from 'classnames';
 import {useRef} from 'react';
@@ -42,33 +42,19 @@ export const dockView = hoistCmp.factory({
                 item: div({className: 'xh-dock-view__body', item: elementFromContent(model.content)})
             });
 
-        // 1) Render collapsed
-        if (collapsed) {
-            return vbox({
-                width: collapsedWidth,
-                className: classNames(className, 'xh-dock-view--collapsed'),
-                items: [header, unmount ? null : body]
-            });
-        }
 
-        // 1) Render docked
-        if (docked) {
-            return vbox({
-                width,
-                height,
-                className: classNames(className, 'xh-dock-view--docked'),
-                items: [header, unmount ? null : body]
-            });
-        }
+        const suffix = collapsed ? 'collapsed' : docked ? 'docked' : 'dialog';
 
-        // 2) Render in Dialog
-        return dialog({
-            className: classNames(className, 'xh-dock-view--dialog'),
-            style: {width, height},
-            isOpen: true,
-            onClose: () => model.onClose(),
+        return fullScreenSupport({
             canOutsideClickClose: false,
-            items: [header, body]
+            onClose: () => model.onClose(),
+            style: {width, height},
+            item: vbox({
+                width: collapsed ? collapsedWidth : width,
+                height,
+                className: classNames(className, `xh-dock-view--${suffix}`),
+                items: [header, unmount ? null : body]
+            })
         });
     }
 });